pxa910: spi: null_dma_buf out of allocated memory.
spi_alloc_master allocates memory for null_dma_buf,
but actually null_dma_buf is out of that region.
The null_dma_buf address may be used by others.
When doing spi write, null_dma_buf will be written,
and impacts that using the address.

+++ drivers/spi/pxa2xx_spi.c
@@ -1569,7 +1569,7 @@ static int __devinit pxa2xx_spi_probe(struct platform_device *pdev)
 	master->transfer = transfer;
 
 	drv_data->ssp_type = ssp->type;
-	drv_data->null_dma_buf = (u32 *)ALIGN((u32)(drv_data +
+	drv_data->null_dma_buf = (u32 *)ALIGN((u32)((u32)drv_data +
 						sizeof(struct driver_data)), 8);
 
 	drv_data->ioaddr = ssp->mmio_base;
